fix: every bug found in MR + deleted all print statements

+++ src/avonic_camera_api/converter.py
@@ -25,7 +25,7 @@ def vector_angle(v: np.array) -> (float, float):
     Returns:
         alpha - horizontal angle, beta - vertical angle in rad
     """
-    if len(v) != 3 or not type(v[0]) is float:
+    if len(v) != 3 or not type(v[0]) is np.float64:
         raise TypeError("vector must contain three floats")
     norm = np.linalg.norm(v)
     if norm == 0:
